Allow configuration of debug flags
Add parameters to set the debug attribute of the client and service nodes. Note
that the values are no longer boolean but one of TRACE, INFO, WARN, ERROR,
SEVERE or OFF. This is the debug logging going to syslog.
Since the (current) code automatically but slightly later (re)sets the debug
attribute for the logging node, which configures the logging to file, we set the
debug node directly to the corresponding value: false for OFF, true for all
others.