Properly handle empty scopes
Explicitly check whether a scope is set when using them in the cfg.xml files and
skip empty ones in the loops. This is particularly important since we set a
default mp_server_sshkey_scope, hence if we override it in local configs, it is
still defined as variable. It's easiest to handle this in the cfg.xml files.

Additionally add template config for overriding the mp_server_sshkey_scope
and the enabling of the autoregistration endpoint.