mariadb 5.5 is picky about TIMESTAMP columns
In MySQL and mariaDB 5.5, a TIMESTAMP column with a DEFAULT must be the first
appearing TIMESTAMP column or we get the error:
    "Incorrect table definition; there can be only one TIMESTAMP column with
    CURRENT_TIMESTAMP in DEFAULT or ON UPDATE clause"
Moving the creation_ts column up fixes this issue. Alternatively, we probably
could also give the issuedat column a default 0.
See also https://dev.mysql.com/doc/refman/5.5/en/timestamp-initialization.html
and https://stackoverflow.com/questions/4489548/why-there-can-be-only-one-timestamp-column-with-current-timestamp-in-default-cla