Add support for multiple vomses configs
We add a new configurable variable 'vomses_input' that can be either a file or
directory under roles/basic/files/. It's default is set in
roles/common/defaults/main.yml and can be overridden in config/*/common_env.yml
to distinguish between different installations.
It will be copied (recursively if a directory) to /etc/vomses on the remote
hosts.